requirements.txt
# FastAPI 및 웹 서비스 관련
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# 문서 처리 라이브러리
unstructured[pdf]==0.11.2
pypdf2==3.0.1
pillow==10.1.0

# 데이터 처리
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.8.0

# 로깅 및 유틸리티
coloredlogs==15.0.1
python-json-logger==2.0.7

# Google Gemini API
google-genai==1.0.0

# 테스트
pytest==7.4.3
httpx==0.25.2 